finalized v1 for submission

+++ backend/src/lambda/http/campgrounds/deleteCampground.ts
@@ -2,7 +2,7 @@ import 'source-map-support/register'
 import { APIGatewayProxyEvent, APIGatewayProxyResult, APIGatewayProxyHandler } from 'aws-lambda'
 import { createLogger } from '../../../utils/logger'
 import {getUsername} from '../../utils'
-import {deleteCampground} from '../../../businessLogic/campgrounds';
+import {deleteCampground, getCampground} from '../../../businessLogic/campgrounds';
 
 const logger = createLogger('deleteCampgrounds')
 
@@ -10,12 +10,34 @@ export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEven
   logger.info("Processing event", event)
   const author = getUsername(event)  
   const campgroundId = event.pathParameters.campgroundId
-  
+  const item = await getCampground(campgroundId)
+  if (!item){
+    return {
+      statusCode: 404,
+      headers: {
+        'Access-Control-Allow-Origin': '*'
+      },
+      body: JSON.stringify({
+        error: 'Not Found',
+        message:'Cannot find campground with given id'
+      })
+    }
+  }
+  if (item.author !== author){
+    return {
+      statusCode: 404,
+      headers: {
+        'Access-Control-Allow-Origin': '*'
+      },
+      body: JSON.stringify({
+        error: 'Author',
+        message:'User is not the author'
+      })
+    }
+  }
+  logger.info('Trying to delete item', {item})
+
   try{
-    logger.info('Trying to delete item', {
-      author: author,
-      campgroundId: campgroundId
-    })
     await deleteCampground(author, campgroundId)
     return {
       statusCode: 200,
@@ -37,4 +59,3 @@ export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEven
     }
   }
 }
-
